+++ android/feature/setting/src/main/kotlin/com/andlife/setting/SettingNavGraph.kt
@@ -6,9 +6,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import androidx.navigation.toRoute
 import com.andlife.setting.screen.SettingRoute
-import com.andlife.setting.screen.WebViewScreen
 import kotlinx.serialization.Serializable
 
 object PolicyUrl {
@@ -19,15 +17,8 @@ object PolicyUrl {
 @Serializable
 data object Setting
 
-@Serializable
-data class WebViewRoute(val url: String, val title: String)
-
 fun NavController.navigateToSetting(navOptions: NavOptions) {
     navigate(Setting, navOptions)
-}
-
-fun NavController.navigateToWebView(url: String, title: String, navOptions: NavOptions) {
-    navigate(WebViewRoute(url, title), navOptions)
 }
 
 fun NavGraphBuilder.settingNavGraph(
@@ -45,19 +36,6 @@ fun NavGraphBuilder.settingNavGraph(
             onNavigateToWebView = onNavigateToWebView,
             onSignedOut = onSignedOut,
             onLogout = onLogout
-        )
-    }
-}
-
-fun NavGraphBuilder.webViewNavGraph(
-    onNavigateBack: () -> Unit,
-) {
-    composable<WebViewRoute> { backStackEntry ->
-        val route = backStackEntry.toRoute<WebViewRoute>()
-        WebViewScreen(
-            url = route.url,
-            title = route.title,
-            onNavigateBack = onNavigateBack,
         )
     }
 }
