+++ android/feature/setting/src/main/res/values/strings.xml
@@ -34,5 +34,4 @@
     <string name="desc_quit">회원탈퇴 상세 아이콘 버튼</string>
     <string name="txt_sign_out">정말 회원 탈퇴를 하시겠습니까?\n회원 탈퇴 시 모든 초대 정보가 삭제됩니다.</string>
     <string name="fail_sign_out">회원 탈퇴에 실패했습니다 다시 시도해 주세요.</string>
-    <string name="txt_error_network">네트워크 연결을 확인해주세요</string>
 </resources>
