+++ android/feature/webview/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("invitation.android.feature")
+}
+
+android {
+    namespace = "com.andlife.webview"
+}
