+++ android/settings.gradle.kts
@@ -45,3 +45,4 @@ include(":core:media")
 include(":feature:editor")
 include(":feature:login")
 include(":feature:setting")
+include(":feature:webview")
